Camera Lerp

+++ src/robatortas/code/files/core/entities/Mob.java
@@ -6,7 +6,7 @@ import robatortas.code.files.project.archive.tileArchive.TileArchive;
 
 public class Mob extends EntityManager {
 
-	public int xa, ya;
+	public float xa, ya;
 	
 	protected int avgHealth = 10;
 	public int health = avgHealth;
@@ -21,16 +21,16 @@ public class Mob extends EntityManager {
 	protected boolean isSwimming = false;
 	public int dir = 3;
 	
-	public void move(int xa, int ya) {
+	public void move(float xa, float ya) {
 		super.move(xa, ya);
 	}
 	
-	public void move2(int xa, int ya) {
+	public void move2(float xa, float ya) {
 		super.move2(xa, ya);
 	}
 	
 	public boolean isSwimming() {
-		if(LevelManager.level.getLevel(x >> 4, y >> 4) == TileArchive.water) isSwimming = true;
+		if(LevelManager.level.getLevel((int)x >> 4, (int)y >> 4) == TileArchive.water) isSwimming = true;
 		else isSwimming = false;
 		return isSwimming;
 	}
